# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "RiskSense AI API",
    "version": "0.1.0",
    "description": "Generated from Zod schemas. See docs/ai/API.md."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "sessionId": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Session-Id"
      }
    },
    "schemas": {
      "ErrorEnvelope": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {
                "nullable": true
              }
            },
            "required": [
              "code",
              "message"
            ]
          },
          "meta": {
            "type": "object",
            "properties": {
              "requestId": {
                "type": "string"
              }
            },
            "required": [
              "requestId"
            ]
          }
        },
        "required": [
          "error",
          "meta"
        ]
      },
...
      "Scenario": {
        "type": "object",
        "additionalProperties": {
          "nullable": true
        }
      },
...
    },
...
  },
  "paths": {
...
    "/scenarios": {
      "get": {
        "security": [
          {
            "bearerAuth": [],
            "sessionId": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "personaKey",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "draft",
                "active",
                "deactivated"
              ]
            },
            "required": false,
            "name": "status",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "key",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "current",
                "all"
              ]
            },
            "required": false,
            "name": "view",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "List Scenarios",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Scenario"
                      }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "requestId"
                      ]
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "security": [
          {
            "bearerAuth": [],
            "sessionId": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "key": {
                    "type": "string",
                    "pattern": "^[a-z][a-z0-9_]{1,63}$"
                  },
                  "personaKey": {
                    "type": "string",
                    "pattern": "^[a-z][a-z0-9_]{1,63}$"
                  },
                  "name": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 160
                  },
                  "description": {
                    "type": "string",
                    "minLength": 10,
                    "maxLength": 4000
                  },
                  "businessContext": {
                    "type": "string",
                    "minLength": 5,
                    "maxLength": 4000
                  },
                  "learningObjective": {
                    "type": "string",
                    "maxLength": 2000
                  },
                  "riskIndicators": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "minLength": 1
                    },
                    "default": []
                  },
                  "conversationFlow": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "questionKey": {
                          "type": "string",
                          "pattern": "^[a-z][a-z0-9_]{1,63}$"
                        },
                        "showIf": {
                          "type": "object",
                          "properties": {
                            "factKey": {
                              "type": "string",
                              "pattern": "^[a-z][a-z0-9_]{1,63}$"
                            },
                            "equals": {
                              "anyOf": [
                                {
                                  "type": "string"
                                },
                                {
                                  "type": "number"
                                },
                                {
                                  "type": "boolean"
                                }
                              ]
                            }
                          },
                          "required": [
                            "factKey",
                            "equals"
                          ]
                        }
                      },
                      "required": [
                        "questionKey"
                      ]
                    },
                    "default": []
                  },
                  "requiredFactKeys": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "pattern": "^[a-z][a-z0-9_]{1,63}$"
                    },
                    "default": []
                  },
                  "expectedClassification": {
                    "type": "string",
                    "enum": [
                      "monitor_only",
                      "risk",
                      "elevated_risk",
                      "issue"
                    ]
                  },
                  "reasoningExample": {
                    "type": "string",
                    "maxLength": 4000
                  },
                  "recommendedActions": {
                    "type": "object",
                    "properties": {
                      "monitor_only": {
                        "type": "object",
                        "properties": {
                          "decisionRecommendation": {
                            "type": "string",
                            "minLength": 2
                          },
                          "nextSteps": {
                            "type": "array",
                            "items": {
                              "type": "string",
                              "minLength": 1
                            },
                            "default": []
                          }
                        },
                        "required": [
                          "decisionRecommendation"
                        ]
                      },
                      "risk": {
                        "type": "object",
                        "properties": {
                          "decisionRecommendation": {
                            "type": "string",
                            "minLength": 2
                          },
                          "nextSteps": {
                            "type": "array",
                            "items": {
                              "type": "string",
                              "minLength": 1
                            },
                            "default": []
                          }
                        },
                        "required": [
                          "decisionRecommendation"
                        ]
                      },
                      "elevated_risk": {
                        "type": "object",
                        "properties": {
                          "decisionRecommendation": {
                            "type": "string",
                            "minLength": 2
                          },
                          "nextSteps": {
                            "type": "array",
                            "items": {
                              "type": "string",
                              "minLength": 1
                            },
                            "default": []
                          }
                        },
                        "required": [
                          "decisionRecommendation"
                        ]
                      },
                      "issue": {
                        "type": "object",
                        "properties": {
                          "decisionRecommendation": {
                            "type": "string",
                            "minLength": 2
                          },
                          "nextSteps": {
                            "type": "array",
                            "items": {
                              "type": "string",
                              "minLength": 1
                            },
                            "default": []
                          }
                        },
                        "required": [
                          "decisionRecommendation"
                        ]
                      }
                    },
                    "default": {}
                  }
                },
                "required": [
                  "key",
                  "personaKey",
                  "name",
                  "description",
                  "businessContext"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Scenario created (draft)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Scenario"
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "requestId"
                      ]
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ]
                }
              }
            }
          },
          "409": {
            "description": "CONFLICT (duplicate key)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/scenarios/{id}": {
      "get": {
        "security": [
          {
            "bearerAuth": [],
            "sessionId": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Scenario",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Scenario"
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "requestId"
                      ]
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ]
                }
              }
            }
          }
        }
      },
      "patch": {
        "security": [
          {
            "bearerAuth": [],
            "sessionId": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "personaKey": {
                    "type": "string",
                    "pattern": "^[a-z][a-z0-9_]{1,63}$"
                  },
                  "name": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 160
                  },
                  "description": {
                    "type": "string",
                    "minLength": 10,
                    "maxLength": 4000
                  },
                  "businessContext": {
                    "type": "string",
                    "minLength": 5,
                    "maxLength": 4000
                  },
                  "learningObjective": {
                    "type": "string",
                    "maxLength": 2000
                  },
                  "riskIndicators": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "minLength": 1
                    },
                    "default": []
                  },
                  "conversationFlow": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "questionKey": {
                          "type": "string",
                          "pattern": "^[a-z][a-z0-9_]{1,63}$"
                        },
                        "showIf": {
                          "type": "object",
                          "properties": {
                            "factKey": {
                              "type": "string",
                              "pattern": "^[a-z][a-z0-9_]{1,63}$"
                            },
                            "equals": {
                              "anyOf": [
                                {
                                  "type": "string"
                                },
                                {
                                  "type": "number"
                                },
                                {
                                  "type": "boolean"
                                }
                              ]
                            }
                          },
                          "required": [
                            "factKey",
                            "equals"
                          ]
                        }
                      },
                      "required": [
                        "questionKey"
                      ]
                    },
                    "default": []
                  },
                  "requiredFactKeys": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "pattern": "^[a-z][a-z0-9_]{1,63}$"
                    },
                    "default": []
                  },
                  "expectedClassification": {
                    "type": "string",
                    "enum": [
                      "monitor_only",
                      "risk",
                      "elevated_risk",
                      "issue"
                    ]
                  },
                  "reasoningExample": {
                    "type": "string",
                    "maxLength": 4000
                  },
                  "recommendedActions": {
                    "type": "object",
                    "properties": {
                      "monitor_only": {
                        "type": "object",
                        "properties": {
                          "decisionRecommendation": {
                            "type": "string",
                            "minLength": 2
                          },
                          "nextSteps": {
                            "type": "array",
                            "items": {
                              "type": "string",
                              "minLength": 1
                            },
                            "default": []
                          }
                        },
                        "required": [
                          "decisionRecommendation"
                        ]
                      },
                      "risk": {
                        "type": "object",
                        "properties": {
                          "decisionRecommendation": {
                            "type": "string",
                            "minLength": 2
                          },
                          "nextSteps": {
                            "type": "array",
                            "items": {
                              "type": "string",
                              "minLength": 1
                            },
                            "default": []
                          }
                        },
                        "required": [
                          "decisionRecommendation"
                        ]
                      },
                      "elevated_risk": {
                        "type": "object",
                        "properties": {
                          "decisionRecommendation": {
                            "type": "string",
                            "minLength": 2
                          },
                          "nextSteps": {
                            "type": "array",
                            "items": {
                              "type": "string",
                              "minLength": 1
                            },
                            "default": []
                          }
                        },
                        "required": [
                          "decisionRecommendation"
                        ]
                      },
                      "issue": {
                        "type": "object",
                        "properties": {
                          "decisionRecommendation": {
                            "type": "string",
                            "minLength": 2
                          },
                          "nextSteps": {
                            "type": "array",
                            "items": {
                              "type": "string",
                              "minLength": 1
                            },
                            "default": []
                          }
                        },
                        "required": [
                          "decisionRecommendation"
                        ]
                      }
                    },
                    "default": {}
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Draft updated in place, or a new draft version created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Scenario"
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "requestId"
                      ]
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/scenarios/{id}/history": {
      "get": {
        "security": [
          {
            "bearerAuth": [],
            "sessionId": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "All versions, newest first",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Scenario"
                      }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "requestId"
                      ]
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/scenarios/{id}/activate": {
      "post": {
        "security": [
          {
            "bearerAuth": [],
            "sessionId": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Activated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Scenario"
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "requestId"
                      ]
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ]
                }
              }
            }
          },
          "422": {
            "description": "NO_LINKED_QUESTIONS / validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/scenarios/{id}/deactivate": {
      "post": {
        "security": [
          {
            "bearerAuth": [],
            "sessionId": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Deactivated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Scenario"
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "requestId"
                      ]
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  }
}